Split form_transect.rs into subdirectory (908 -> ~400 lines each)
Split the monolithic form_transect.rs into a form_transect/ directory with
three focused modules:
- types.rs (403 lines): TransectZone enum and form-based code constants
- grid.rs (433 lines): TransectGrid resource, FAR-to-level conversion, Saveable
- systems.rs (73 lines): enforce_transect_constraints system and plugin

All public items re-exported from mod.rs so external imports remain unchanged.

Part of #1389
